refactor(pages): added changes to 'Login' and 'Signup' pages
