+++ packages/frontend/src/pages/signup.tsx
@@ -6,12 +6,13 @@ import { useForm } from 'react-hook-form';
 
 import { ReactElement } from 'react';
 import { Button } from '@/components/ui';
-import FormField from '@/components/auth/FormField/FormField';
+import FormField from '@/components/common/FormField/FormField';
 import { signup } from '@/api';
 import UseAlertStore from '@/store/store';
 import { NextPageWithLayout } from './_app';
 import Layout from '@/layouts/Layout';
 
+// TODO: Refactor page + components
 const Signup: NextPageWithLayout = () => {
   const setAlert = UseAlertStore((state) => state.setAlert);
 
@@ -21,7 +22,15 @@ const Signup: NextPageWithLayout = () => {
     register,
     handleSubmit,
     formState: { errors, isValid, isSubmitting },
-  } = useForm();
+  } = useForm({
+    defaultValues: {
+      firstName: '',
+      lastName: '',
+      username: '',
+      email: '',
+      password: '',
+    },
+  });
 
   const NewUserMutation = useMutation({
     mutationFn: signup,
@@ -44,7 +53,7 @@ const Signup: NextPageWithLayout = () => {
   };
 
   return (
-    <section className="h-screen px-4 flex flex-col justify-center items-center">
+    <section className="flex flex-col items-center justify-center h-screen px-4">
       <article className="w-full max-w-[500px] mx-auto p-10 border-[1px] rounded-lg text-sm">
         <div className="mb-8">
           <h1 className="text-2xl font-bold">Create an account</h1>
@@ -56,7 +65,7 @@ const Signup: NextPageWithLayout = () => {
             placeholder="Enter your firstname"
             register={register}
             options={{ required: 'First name is required' }}
-            error={errors}
+            error={errors.firstName}
           />
           <FormField
             id="lastName"
@@ -64,7 +73,7 @@ const Signup: NextPageWithLayout = () => {
             placeholder="Enter your lastname"
             register={register}
             options={{ required: 'Last name is required' }}
-            error={errors}
+            error={errors.lastName}
           />
           <FormField
             id="username"
@@ -78,7 +87,7 @@ const Signup: NextPageWithLayout = () => {
                 message: 'Username should be more than 1 character',
               },
             }}
-            error={errors}
+            error={errors.username}
           />
           <FormField
             id="email"
@@ -93,7 +102,7 @@ const Signup: NextPageWithLayout = () => {
                 message: 'Email is not valid.',
               },
             }}
-            error={errors}
+            error={errors.email}
           />
           <FormField
             id="password"
@@ -108,7 +117,7 @@ const Signup: NextPageWithLayout = () => {
                 message: 'Password should be at-least 6 characters.',
               },
             }}
-            error={errors}
+            error={errors.password}
           />
           <Button
             className="w-full py-4 my-8 text-base font-normal"
@@ -118,10 +127,7 @@ const Signup: NextPageWithLayout = () => {
             Create my account
           </Button>
         </form>
-        <div
-          className="my-5 text-base
-   text-center"
-        >
+        <div className="my-5 text-base text-center">
           <p className="">
             Already got an account?
             <Link
